requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httplib2==0.22.0
hyperframe==6.1.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matrix-nio==0.25.2
multidict==6.2.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-socks==2.7.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.13.0
unpaddedbase64==2.1.0
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
